refactor: Clean utils classes
