+++ backend/server/src/main/java/ai/verta/modeldb/dataset/DatasetServiceImpl.java
@@ -163,7 +163,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
       responseObserver.onCompleted();
 
     } catch (Exception e) {
-      CommonUtils.observeError(responseObserver, e, CreateDataset.Response.getDefaultInstance());
+      CommonUtils.observeError(responseObserver, e);
     }
   }
 
@@ -238,7 +238,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
       responseObserver.onCompleted();
 
     } catch (Exception e) {
-      CommonUtils.observeError(responseObserver, e, GetAllDatasets.Response.getDefaultInstance());
+      CommonUtils.observeError(responseObserver, e);
     }
   }
 
@@ -261,7 +261,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
       responseObserver.onCompleted();
 
     } catch (Exception e) {
-      CommonUtils.observeError(responseObserver, e, DeleteDataset.Response.getDefaultInstance());
+      CommonUtils.observeError(responseObserver, e);
     }
   }
 
@@ -287,7 +287,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
       responseObserver.onCompleted();
 
     } catch (Exception e) {
-      CommonUtils.observeError(responseObserver, e, GetDatasetById.Response.getDefaultInstance());
+      CommonUtils.observeError(responseObserver, e);
     }
   }
 
@@ -308,7 +308,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
       responseObserver.onCompleted();
 
     } catch (Exception e) {
-      CommonUtils.observeError(responseObserver, e, FindDatasets.Response.getDefaultInstance());
+      CommonUtils.observeError(responseObserver, e);
     }
   }
 
@@ -368,7 +368,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
       responseObserver.onCompleted();
 
     } catch (Exception e) {
-      CommonUtils.observeError(responseObserver, e, GetDatasetByName.Response.getDefaultInstance());
+      CommonUtils.observeError(responseObserver, e);
     }
   }
 
@@ -416,7 +416,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
 
     } catch (Exception e) {
       CommonUtils.observeError(
-          responseObserver, e, UpdateDatasetName.Response.getDefaultInstance());
+          responseObserver, e);
     }
   }
 
@@ -461,7 +461,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
 
     } catch (Exception e) {
       CommonUtils.observeError(
-          responseObserver, e, UpdateDatasetDescription.Response.getDefaultInstance());
+          responseObserver, e);
     }
   }
 
@@ -510,7 +510,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
       responseObserver.onCompleted();
 
     } catch (Exception e) {
-      CommonUtils.observeError(responseObserver, e, AddDatasetTags.Response.getDefaultInstance());
+      CommonUtils.observeError(responseObserver, e);
     }
   }
 
@@ -522,7 +522,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
       throw new ModelDBException("Not supported", Code.UNIMPLEMENTED);
 
     } catch (Exception e) {
-      CommonUtils.observeError(responseObserver, e, GetTags.Response.getDefaultInstance());
+      CommonUtils.observeError(responseObserver, e);
     }
   }
 
@@ -577,7 +577,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
 
     } catch (Exception e) {
       CommonUtils.observeError(
-          responseObserver, e, DeleteDatasetTags.Response.getDefaultInstance());
+          responseObserver, e);
     }
   }
 
@@ -635,7 +635,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
 
     } catch (Exception e) {
       CommonUtils.observeError(
-          responseObserver, e, AddDatasetAttributes.Response.getDefaultInstance());
+          responseObserver, e);
     }
   }
 
@@ -690,7 +690,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
 
     } catch (Exception e) {
       CommonUtils.observeError(
-          responseObserver, e, UpdateDatasetAttributes.Response.getDefaultInstance());
+          responseObserver, e);
     }
   }
 
@@ -757,7 +757,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
 
     } catch (Exception e) {
       CommonUtils.observeError(
-          responseObserver, e, DeleteDatasetAttributes.Response.getDefaultInstance());
+          responseObserver, e);
     }
   }
 
@@ -785,7 +785,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
       responseObserver.onCompleted();
 
     } catch (Exception e) {
-      CommonUtils.observeError(responseObserver, e, DeleteDatasets.Response.getDefaultInstance());
+      CommonUtils.observeError(responseObserver, e);
     }
   }
 
@@ -903,7 +903,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
 
     } catch (Exception e) {
       CommonUtils.observeError(
-          responseObserver, e, LastExperimentByDatasetId.Response.getDefaultInstance());
+          responseObserver, e);
     }
   }
 
@@ -975,7 +975,7 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
 
     } catch (Exception e) {
       CommonUtils.observeError(
-          responseObserver, e, GetExperimentRunByDataset.Response.getDefaultInstance());
+          responseObserver, e);
     }
   }
 }
